Some %s was changed to % lu + cast of only one of operands was added to calculations

+++ 09_statvfs/09.1_statvfs.c
@@ -5,6 +5,7 @@
 #include <sys/statvfs.h>
 #include <sys/stat.h>
 #include <unistd.h>
+#include <stdint.h>
 
 int main (int argc, char* argv[]) {
 
@@ -17,12 +18,12 @@ int main (int argc, char* argv[]) {
 
 	printf("File system lock size:                      %lu\n", st_vfs.f_bsize);  /* printing fs info */
 	printf("Fragment size:                              %lu\n", st_vfs.f_frsize);
-	printf("Size of filesystem in fragment size units:  %u\n", st_vfs.f_blocks);
-	printf("Amount of free blocks:                      %u\n", st_vfs.f_bfree);
-	printf("Amount of free blocks for non-root:         %u\n", st_vfs.f_bavail);
-	printf("Amount of inodes:                           %u\n", st_vfs.f_files);
-	printf("Amount of free inodes:                      %u\n", st_vfs.f_ffree);
-	printf("Amount of free inodes for non-root:         %u\n", st_vfs.f_favail);
+	printf("Size of filesystem in fragment size units:  %lu\n", (uintmax_t)st_vfs.f_blocks);
+	printf("Amount of free blocks:                      %lu\n", (uintmax_t)st_vfs.f_bfree);
+	printf("Amount of free blocks for non-root:         %lu\n", (uintmax_t)st_vfs.f_bavail);
+	printf("Amount of inodes:                           %lu\n", (uintmax_t)st_vfs.f_files);
+	printf("Amount of free inodes:                      %lu\n", (uintmax_t)st_vfs.f_ffree);
+	printf("Amount of free inodes for non-root:         %lu\n", (uintmax_t)st_vfs.f_favail);
 	printf("File system ID:                             %lu\n", st_vfs.f_fsid);
 	printf("Mount flags:                                %lu\n", st_vfs.f_flag);
 	printf("Maximum filename length:                    %lu\n", st_vfs.f_namemax);
@@ -37,7 +38,7 @@ int main (int argc, char* argv[]) {
 
 	short res = 1; /* creating variable  which contains result of program work */
 
-	if ((st_vfs.f_favail < 1) || ((long int)(st_vfs.f_bavail * st_vfs.f_bsize) < sb.st_size)) /* checking if there is inode numbers or        */
+	if ((st_vfs.f_favail < 1) || (((unsigned long)st_vfs.f_bavail * st_vfs.f_bsize) < sb.st_size)) /* checking if there is inode numbers or        */
 		res = 0;                                                                                /* if there is enough space for this file in fs */
 
 	printf("Does the file fit:                          %s\n", res ? "Yes" : "No"); /* printing results */
